refactor(kanban): extract workspace/worker mixins from kanban_db.py (shard s4)
