+++ hermes_cli/main.py
@@ -7640,291 +7640,21 @@ def _load_installable_optional_extras(group: str = "all") -> list[str]:
 # packages. Cleared only after import-probe repair confirms healthy (not when
 # probes are unavailable/indeterminate). Narrow lazy probes must NEVER clear
 # the generic core marker (#58004 review).
-def _update_marker_path() -> Path:
-    return PROJECT_ROOT / ".update-incomplete"
-
-
-def _lazy_refresh_marker_path() -> Path:
-    return PROJECT_ROOT / ".lazy-refresh-incomplete"
-
-
-def _pytest_owns_live_checkout(root: Path) -> bool:
-    """True when running under pytest AND ``root`` is this checkout itself.
-
-    Tests that drive update/recovery without sandboxing ``PROJECT_ROOT``
-    must neither litter the live repo root with recovery breadcrumbs
-    (a leftover ``.lazy-refresh-incomplete`` / ``.update-incomplete``
-    false-arms recovery on the developer's next real launch) nor run a real
-    reinstall against the executing venv. Sandboxed tests point at a
-    tmp_path and are unaffected (same posture as
-    ``managed_scope._under_pytest``)."""
-    return (
-        "PYTEST_CURRENT_TEST" in os.environ
-        and root == Path(__file__).resolve().parent.parent
-    )
-
-
-def _clear_marker_file(path: Path, *, label: str) -> None:
-    """Remove an update-recovery breadcrumb. Never raises."""
-    try:
-        path.unlink()
-    except FileNotFoundError:
-        pass
-    except OSError as exc:
-        logger.debug("Could not clear %s marker: %s", label, exc)
-
-
-def _clear_update_incomplete_marker() -> None:
-    """Remove the interrupted core-install breadcrumb. Never raises."""
-    _clear_marker_file(_update_marker_path(), label="update-incomplete")
-
-
-def _clear_lazy_refresh_incomplete_marker() -> None:
-    """Remove the interrupted lazy-refresh breadcrumb. Never raises."""
-    _clear_marker_file(_lazy_refresh_marker_path(), label="lazy-refresh-incomplete")
-
-
-def _recover_from_interrupted_install() -> None:
-    """Finish update work left half-done by a prior ``hermes update``.
-
-    Handles two independent breadcrumbs:
-
-    - ``.update-incomplete`` — core ``.[all]`` install interrupted. Recovers
-      via full quarantined reinstall. Never cleared by the narrow lazy-refresh
-      import probes alone.
-    - ``.lazy-refresh-incomplete`` — lazy-backend refresh may have corrupted
-      packages. Recovers via package-only import probes; cleared only when
-      probes confirm healthy/repaired (indeterminate keeps the marker).
-
-    Never raises: a recovery failure must not block launch.  If it can't
-    self-heal it prints the manual command and leaves the relevant marker so
-    the next launch tries again.
-
-    Concurrency: markers live next to the shared venv, so a gateway start
-    plus a CLI launch (or two profiles starting at once) can both see them.
-    An ``O_EXCL`` lockfile ensures only one process runs recovery; the
-    others skip and let the winner clear markers.
-
-    Output: everything — our status lines AND the streamed pip/uv install
-    (which inherits fd 1) — is routed to stderr.  Launches whose stdout is a
-    protocol stream (``hermes acp`` speaks JSON-RPC on stdout) must never get
-    install noise on stdout.
-    """
-    if _pytest_owns_live_checkout(PROJECT_ROOT):
-        return
-    core_marker = _update_marker_path().exists()
-    lazy_marker = _lazy_refresh_marker_path().exists()
-    if not core_marker and not lazy_marker:
-        return
-
-    # Skip in managed/Docker installs and on PyPI installs with no git checkout:
-    # those don't run the source-tree update path, so a stray marker is not ours
-    # to act on. Just clear it.
-    if not (PROJECT_ROOT / "pyproject.toml").is_file():
-        _clear_update_incomplete_marker()
-        _clear_lazy_refresh_incomplete_marker()
-        return
-
-    # Single-flight guard: atomically claim the recovery lock. If another
-    # process holds it, skip — it is running the same reinstall into the same
-    # shared venv right now. A crashed holder leaves a stale lock; break it
-    # after an hour (well past any realistic install) so recovery can't be
-    # wedged forever.
-    lock_path = PROJECT_ROOT / ".update-incomplete.lock"
-    try:
-        fd = os.open(lock_path, os.O_CREAT | os.O_EXCL | os.O_WRONLY)
-        os.write(fd, f"{os.getpid()}\n".encode())
-        os.close(fd)
-    except FileExistsError:
-        try:
-            if _time.time() - lock_path.stat().st_mtime > 3600:
-                lock_path.unlink()
-        except OSError:
-            pass
-        return
-    except OSError as exc:
-        # Couldn't create the lock (read-only fs, perms). Proceed unlocked —
-        # the install itself will surface the real problem.
-        logger.debug("Could not create install-recovery lock: %s", exc)
-
-    saved_stdout_fd = None
-    saved_sys_stdout = sys.stdout
-    try:
-        # Route Python-level prints AND subprocess-inherited fd 1 to stderr
-        # for the duration of recovery (see docstring: ACP stdout safety).
-        try:
-            saved_stdout_fd = os.dup(1)
-            os.dup2(2, 1)
-        except OSError:
-            saved_stdout_fd = None
-        sys.stdout = sys.stderr
-
-        if lazy_marker:
-            _recover_lazy_refresh_marker_locked()
-
-        if _update_marker_path().exists():
-            _recover_core_update_marker_locked()
-    finally:
-        sys.stdout = saved_sys_stdout
-        if saved_stdout_fd is not None:
-            try:
-                os.dup2(saved_stdout_fd, 1)
-                os.close(saved_stdout_fd)
-            except OSError:
-                pass
-        try:
-            lock_path.unlink()
-        except OSError:
-            pass
-
-
-def _recover_lazy_refresh_marker_locked() -> None:
-    """Heal ``.lazy-refresh-incomplete`` via confirmed import-probe repair."""
-    print(
-        "⚠ A previous lazy-backend refresh may have left the venv unhealthy — "
-        "running import-based package repair..."
-    )
-    install_prefix, install_env = _default_venv_install_target()
-    status = _repair_venv_via_import_probes(install_prefix, env=install_env)
-    if status in ("healthy", "repaired"):
-        _clear_lazy_refresh_incomplete_marker()
-        print("✓ Lazy-refresh venv recovery confirmed — install is healthy again.")
-        return
-    if status == "indeterminate":
-        print(
-            "  ⚠ Import probes unavailable — cannot confirm venv health. "
-            "Leaving `.lazy-refresh-incomplete` for the next launch."
-        )
-    else:
-        print(
-            "  ⚠ Lazy-refresh package repair incomplete. "
-            "Leaving `.lazy-refresh-incomplete` for the next launch."
-        )
-        print("  Recover manually with:")
-        all_specs = _lazy_refresh_repair_specs(
-            sorted(set(_LAZY_REFRESH_REPAIR_PACKAGES.values()))
-        )
-        print(
-            f"    {' '.join(install_prefix)} install --force-reinstall "
-            + " ".join(shlex.quote(s) for s in all_specs)
-        )
-
-
-def _recover_core_update_marker_locked() -> None:
-    """Heal ``.update-incomplete`` via full ``.[all]`` reinstall only.
-
-    Narrow lazy-refresh import probes are not sufficient proof that a generic
-    interrupted core install finished — a missing dep outside that probe set
-    would otherwise look healthy and clear the breadcrumb too early.
-    """
-    print(
-        "⚠ A previous `hermes update` was interrupted mid-install — "
-        "finishing dependency installation now..."
-    )
-
-    # Windows: a normal ``hermes.exe`` launch always has the launcher as an
-    # ancestor. Full editable reinstall uses quarantine so the live shim can
-    # still be replaced. Package-only import repair may help as first aid but
-    # must NEVER clear this core marker on its own (#58004 review).
-    self_locked = _windows_running_hermes_launcher_locked()
-    if self_locked:
-        install_prefix, install_env = _default_venv_install_target()
-        print(
-            "  → Running from hermes.exe; applying package-only first aid, "
-            "then quarantined full reinstall (core marker stays until that "
-            "succeeds)..."
-        )
-        _repair_venv_via_import_probes(install_prefix, env=install_env)
-
-    try:
-        from hermes_cli.managed_uv import ensure_uv
-
-        # Always bootstrap pip first: a killed install can leave the venv with
-        # no pip module at all, and uv may also be gone. ensurepip restores a
-        # known-good pip so at least the plain-pip path below can proceed.
-        try:
-            subprocess.run(
-                [sys.executable, "-m", "ensurepip", "--upgrade", "--default-pip"],
-                cwd=PROJECT_ROOT,
-                capture_output=True,
-            )
-        except Exception as exc:
-            logger.debug("ensurepip during install recovery failed: %s", exc)
-
-        uv_bin = ensure_uv()
-        if uv_bin:
-            uv_env = {**os.environ, "VIRTUAL_ENV": str(PROJECT_ROOT / "venv")}
-            if _is_termux_env(uv_env):
-                uv_env.pop("PYTHONPATH", None)
-                uv_env.pop("PYTHONHOME", None)
-            _install_python_dependencies_with_optional_fallback(
-                [uv_bin, "pip"],
-                env=uv_env,
-                group="termux-all" if _is_termux_env(uv_env) else "all",
-            )
-        else:
-            _install_python_dependencies_with_optional_fallback(
-                [sys.executable, "-m", "pip"],
-                group="termux-all" if _is_termux_env() else "all",
-            )
-
-        _clear_update_incomplete_marker()
-        print("✓ Dependency installation recovered — your install is healthy again.")
-    except Exception as exc:
-        # Leave the marker in place so the next launch retries. Give the user
-        # the exact manual recovery command in the meantime.
-        logger.debug("Interrupted-install recovery failed: %s", exc)
-        print("✗ Could not auto-recover the interrupted install.")
-        if self_locked:
-            print(
-                "  Hermes is still running from the launcher that needs "
-                "replacing. Close other Hermes windows, restart from a "
-                "different terminal, then run:"
-            )
-            print(f'    cd /d "{PROJECT_ROOT}"')
-            print(
-                f'    "{sys.executable}" -m pip install -e ".[all]"'
-            )
-        else:
-            print("  Recover manually with:")
-            print(f"    cd {PROJECT_ROOT}")
-            print(f"    {sys.executable} -m ensurepip --upgrade")
-            print(f"    {sys.executable} -m pip install -e '.[all]'")
-
-
-def _windows_running_hermes_launcher_locked() -> bool:
-    """True when a venv ``hermes*.exe`` shim is this process or an ancestor.
-
-    Best-effort: returns False when psutil is unavailable or inspection fails.
-    """
-    if not _is_windows():
-        return False
-    scripts_dir = _venv_scripts_dir()
-    if scripts_dir is None:
-        return False
-    shims = _hermes_exe_shims(scripts_dir)
-    if not shims:
-        return False
-    shim_set: set[str] = set()
-    for shim in shims:
-        try:
-            shim_set.add(str(shim.resolve()).lower())
-        except OSError:
-            shim_set.add(str(shim).lower())
-    try:
-        import psutil
-
-        me = psutil.Process()
-        for proc in [me] + list(me.parents()):
-            try:
-                exe_norm = str(Path(proc.exe()).resolve()).lower()
-            except Exception:
-                continue
-            if exe_norm in shim_set:
-                return True
-    except Exception:
-        return False
-    return False
+# Update-recovery breadcrumb helpers extracted to hermes_cli/update_recovery.py
+# (main.py decomposition, mechanical move). Re-exported so callers and test
+# monkeypatches on hermes_cli.main.<name> keep resolving unchanged.
+from hermes_cli.update_recovery import (  # noqa: F401
+    _clear_lazy_refresh_incomplete_marker,
+    _clear_marker_file,
+    _clear_update_incomplete_marker,
+    _lazy_refresh_marker_path,
+    _pytest_owns_live_checkout,
+    _recover_core_update_marker_locked,
+    _recover_from_interrupted_install,
+    _recover_lazy_refresh_marker_locked,
+    _update_marker_path,
+    _windows_running_hermes_launcher_locked,
+)
 
 
 def _default_venv_install_target() -> tuple[list[str], dict[str, str] | None]:
@@ -7983,241 +7713,19 @@ def _run_install_with_heartbeat(
         t.join(timeout=0.2)
 
 
-def _is_windows() -> bool:
-    return sys.platform == "win32"
-
-
-def _venv_scripts_dir() -> Path | None:
-    """Return the venv Scripts directory if we're running inside the project venv."""
-    venv_dir = PROJECT_ROOT / "venv"
-    if not venv_dir.is_dir():
-        return None
-    from hermes_constants import venv_bin_dir
-
-    scripts = venv_bin_dir(venv_dir, windows=_is_windows())
-    return scripts if scripts.is_dir() else None
-
-
-def _hermes_exe_shims(scripts_dir: Path) -> list[Path]:
-    """Entry-point shims that uv may try to rewrite during ``pip install -e .``.
-
-    On Windows these are .exe launchers generated by setuptools/uv. On POSIX
-    they're regular Python scripts which can be replaced atomically — no
-    self-replacement hazard exists outside Windows.
-    """
-    if not _is_windows():
-        return []
-
-    names = set(_load_console_script_names()) or {"hermes", "hermes-agent", "hermes-acp"}
-    # The gateway shim is not a [project.scripts] entry point, but older
-    # update/install paths still rewrite and quarantine it.
-    names.add("hermes-gateway")
-    return [scripts_dir / f"{name}.exe" for name in sorted(names)]
-
-
-def _quarantine_running_hermes_exe(
-    scripts_dir: Path, *, max_attempts: int = 4
-) -> list[tuple[Path, Path]]:
-    """Pre-empt Windows file lock on the running ``hermes.exe``.
-
-    Windows allows RENAMING a mapped/running executable (the kernel tracks the
-    file by handle, not path), but blocks DELETE/REPLACE while it's loaded. uv
-    needs to overwrite the entry-point shims during ``pip install -e .``;
-    when ``hermes update`` runs, ``hermes.exe`` IS the live process, and uv
-    fails with ``Access is denied. (os error 5)``.
-
-    We rename live shims to ``hermes.exe.old.<unix-ms>`` first. uv then writes
-    fresh shims at the original paths. The ``.old`` files are cleaned up on
-    the next hermes invocation by ``_cleanup_quarantined_exes``.
-
-    Rename can still fail when *another* process has opened the .exe without
-    ``FILE_SHARE_DELETE`` — typically AV real-time scanners with transient
-    handles (recovers in <1s), or the Hermes Desktop backend child process
-    (won't recover until the user closes it). We mitigate:
-
-    1. Retry up to ``max_attempts`` times with exponential backoff
-       (100/250/500/1000 ms). Handles the AV-scanner case.
-    2. If all retries fail, schedule the .exe for replacement on next
-       reboot via ``MoveFileExW(MOVEFILE_DELAY_UNTIL_REBOOT)``. This still
-       lets uv create a fresh shim at the original path (Windows will keep
-       the old file's content under a new name until the reboot), so the
-       update can complete; the user just needs to reboot to fully unload
-       the stale image.
-    3. Print a clear warning naming the most likely culprit (running
-       Hermes Desktop / gateway / REPL) and pointing to ``--force``.
-
-    Returns the list of (original, quarantined) pairs so the caller can roll
-    back if the install itself fails before uv writes a replacement. Pairs
-    where we used ``MOVEFILE_DELAY_UNTIL_REBOOT`` are NOT returned — they
-    are already deferred and roll-back is meaningless.
-    """
-    moved: list[tuple[Path, Path]] = []
-    if not _is_windows():
-        return moved
-
-    import time
-
-    stamp = int(time.time() * 1000)
-    # Backoff schedule: first attempt is immediate, subsequent ones sleep.
-    # 100ms / 250ms / 500ms covers the typical AV scanner re-scan window.
-    backoff_ms = [0, 100, 250, 500, 1000]
-    attempts = max(1, min(max_attempts, len(backoff_ms)))
-
-    for shim in _hermes_exe_shims(scripts_dir):
-        if not shim.exists():
-            continue
-        target = shim.with_suffix(shim.suffix + f".old.{stamp}")
-
-        last_exc: OSError | None = None
-        for attempt in range(attempts):
-            delay = backoff_ms[attempt] / 1000.0
-            if delay:
-                time.sleep(delay)
-            try:
-                shim.rename(target)
-                moved.append((shim, target))
-                last_exc = None
-                break
-            except OSError as e:
-                last_exc = e
-                continue
-
-        if last_exc is None:
-            continue
-
-        # All in-process renames failed. Try MoveFileEx with
-        # MOVEFILE_DELAY_UNTIL_REBOOT as a last resort. This succeeds in the
-        # exact case where the inline rename failed (another process holds
-        # the handle without share-delete), at the cost of requiring a
-        # reboot to fully reclaim the old .exe.
-        scheduled = _schedule_replace_on_reboot(shim, target)
-        if scheduled:
-            print(
-                f"  ⚠ {shim.name} is locked by another process; scheduled "
-                f"replacement on next reboot."
-            )
-            print(
-                "    The new shim was written at the same path, but a "
-                "reboot is needed to fully unload the old one."
-            )
-            # Do NOT append to ``moved``: we don't want roll-back to undo a
-            # reboot-deferred operation.
-            continue
-
-        # Truly couldn't budge the .exe. Print an actionable warning and let
-        # uv try its luck — sometimes uv's own retry handling pulls through.
-        print(
-            f"  ⚠ Could not quarantine {shim.name} ({last_exc.__class__.__name__}: "
-            f"another process is holding it open)."
-        )
-        print(
-            "    Close Hermes Desktop, exit other `hermes` REPLs, stop the "
-            "gateway, or pause AV scanning, then re-run `hermes update`."
-        )
-
-    return moved
-
-
-def _schedule_replace_on_reboot(shim: Path, quarantine_target: Path) -> bool:
-    """Schedule ``shim`` -> ``quarantine_target`` via PendingFileRenameOperations.
-
-    Uses Win32 ``MoveFileExW`` with ``MOVEFILE_REPLACE_EXISTING |
-    MOVEFILE_DELAY_UNTIL_REBOOT``. The OS persists the rename in
-    ``HKLM\\System\\CurrentControlSet\\Control\\Session Manager\\
-    PendingFileRenameOperations`` and applies it before any user-mode code
-    runs on next boot — at which point no process can hold the .exe.
-
-    Returns ``True`` if the schedule call succeeded, ``False`` otherwise
-    (non-Windows, ctypes failure, lack of privilege, etc.). Never raises.
-    """
-    if not _is_windows():
-        return False
-    try:
-        import ctypes
-        from ctypes import wintypes
-
-        MOVEFILE_REPLACE_EXISTING = 0x1
-        MOVEFILE_DELAY_UNTIL_REBOOT = 0x4
-
-        MoveFileExW = ctypes.windll.kernel32.MoveFileExW
-        MoveFileExW.argtypes = [wintypes.LPCWSTR, wintypes.LPCWSTR, wintypes.DWORD]
-        MoveFileExW.restype = wintypes.BOOL
-
-        ok = MoveFileExW(
-            str(shim),
-            str(quarantine_target),
-            MOVEFILE_REPLACE_EXISTING | MOVEFILE_DELAY_UNTIL_REBOOT,
-        )
-        return bool(ok)
-    except Exception:
-        return False
-
-
-def _restore_quarantined_exes(moved: list[tuple[Path, Path]]) -> None:
-    """Roll back ``_quarantine_running_hermes_exe`` if uv didn't write replacements."""
-    for original, quarantined in moved:
-        try:
-            if not original.exists() and quarantined.exists():
-                quarantined.rename(original)
-        except OSError:
-            pass
-
-
-def _run_quarantined_install(
-    cmd: list[str],
-    *,
-    env: dict[str, str] | None = None,
-    scripts_dir: Path | None = None,
-) -> None:
-    """Run an editable install, quarantining the running ``hermes.exe`` first.
-
-    Any ``pip install -e .`` (or ``--reinstall``) rewrites the entry-point
-    shims, and on Windows the live ``hermes.exe`` is the running process —
-    pip can neither delete nor overwrite it, so without quarantine the shim
-    is left missing and ``hermes`` drops off PATH. This wraps
-    :func:`_run_install_with_heartbeat` with the same rename-out-of-the-way /
-    restore-on-failure dance that the primary install path uses, so EVERY
-    install that touches the shims is protected — including the
-    verification-repair reinstalls in
-    :func:`_verify_core_dependencies_installed`, which previously called
-    ``_run_install_with_heartbeat`` directly and bypassed quarantine.
-
-    Off-Windows (``scripts_dir is None``) this is a thin pass-through.
-    """
-    moved: list[tuple[Path, Path]] = []
-    if scripts_dir is not None:
-        moved = _quarantine_running_hermes_exe(scripts_dir)
-    try:
-        _run_install_with_heartbeat(cmd, env=env)
-    except BaseException:
-        # Restore shims if pip/uv didn't write replacements (e.g. install
-        # failed before the entry-points step). Don't swallow the error.
-        if scripts_dir is not None:
-            _restore_quarantined_exes(moved)
-        raise
-
-
-def _cleanup_quarantined_exes(scripts_dir: Path | None = None) -> None:
-    """Sweep ``hermes.exe.old.*`` left by prior updates.
-
-    Called early on every hermes invocation. The .old files are unlocked once
-    their owning process exited, so deletion succeeds the next run. Silent
-    no-op when nothing's there or on file-locked / permission errors.
-    """
-    if not _is_windows():
-        return
-    if scripts_dir is None:
-        scripts_dir = _venv_scripts_dir()
-    if scripts_dir is None:
-        return
-    try:
-        for stale in scripts_dir.glob("*.exe.old.*"):
-            try:
-                stale.unlink()
-            except OSError:
-                pass  # still locked or in use — try again next run
-    except OSError:
-        pass
+# Windows hermes.exe quarantine helpers extracted to hermes_cli/win_quarantine.py
+# (main.py decomposition, mechanical move). Re-exported so callers and test
+# monkeypatches on hermes_cli.main.<name> keep resolving unchanged.
+from hermes_cli.win_quarantine import (  # noqa: F401
+    _cleanup_quarantined_exes,
+    _hermes_exe_shims,
+    _is_windows,
+    _quarantine_running_hermes_exe,
+    _restore_quarantined_exes,
+    _run_quarantined_install,
+    _schedule_replace_on_reboot,
+    _venv_scripts_dir,
+)
 
 
 # Import probes for venv corruption after a failed lazy ``uv pip install``.
